+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Make dice roller completely optional by disabling it in the settings
 + Auto adjust size of quick roll input
 + React to player role changes without needing to reload site
++ Add self-roll button to quickrolls
 
 # 1.6.2